eth to dai 2

+++ contracts/interfaces/IWAVAX.sol
@@ -2,9 +2,13 @@
 pragma solidity 0.8.7;
 
 interface IWAVAX {
-    function deposit() external payable;
-    function transfer(address to, uint value) external returns (bool);
-    function balanceOf(address owner) external view returns (uint);
-    function withdraw(uint) external;
-    function approve(address to, uint value) external returns (bool);
+	function deposit() external payable;
+
+	function transfer(address to, uint256 value) external returns (bool);
+
+	function balanceOf(address owner) external view returns (uint256);
+
+	function withdraw(uint256) external;
+
+	function approve(address to, uint256 value) external returns (bool);
 }
